Extract reusable validation for sources.
